Honor PORTAGE_SSH_USER when using Archive class.
* If /etc/ssh/ssh_config does not have a "packages" entry, then the SCP
copy will fail due to an incomplete configuration. But since,
PORTAGE_SSH_USER environment variable has been introduced to be able to
have a more dynamic behavior regarding "packages" access. So we also
check if PORTAGE_SSH_USER is defined and it has a higher priority than
config done in /etc/ssh/ssh_config to be as close as possible to the
portage behavior.